Added script and workflow to check flink version compatiblity in docs
The jbang script will pull the flink-kubernetes-operator-api jar for a
given operator version and check the FlinkVersion enum to see if the
specified flink version is supported and/or deprecated. The GH Action
will then check all operator versions listed in the docs and ensure the
flink version specified in the parent pom is compatible with them.

Signed-off-by: Thomas Cooper <code@tomcooper.dev>